requirements.txt
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
contentful==1.9.0
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
pytest-flask==0.10.0
python-dateutil==2.7.3
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
